+++ packages/navigation/src/theme/types.js
@@ -1,5 +1,5 @@
 // @flow
-export type Color = string;
+export type Color = string | () => string;
 export type Background = Color;
 export type Text = Color;
 export type Line = Color;
@@ -38,6 +38,7 @@ export type Provided = {|
   keyline: Line,
   item: ItemTheme,
   dropdown: ItemTheme,
+  hasDarkmode?: boolean,
 |}
 
 export type RootTheme = {|